Split Sentry client filters into modules
