+++ scripts/debt-ratchet-baseline.json
@@ -4,7 +4,7 @@
   "counts": {
     "any_casts": 4,
     "todo_comments": 0,
-    "oversized_files": 89,
+    "oversized_files": 88,
     "bootstrap_imports": 0,
     "redux_imports": 0,
     "pages_router_files": 0
@@ -97,7 +97,6 @@
     "lib/profile-cms/builder/agent.ts",
     "lib/profile-cms/protocol/v1/validation.ts",
     "lib/twitter/fetcher.ts",
-    "services/api/wave-drops-v2-api.ts",
-    "utils/sentry-client-filters.ts"
+    "services/api/wave-drops-v2-api.ts"
   ]
 }
